ADC is shown on the LCD. The output is wrong though. Only 3 digits are shown.

+++ TouchGFX/gui/include/gui/screen1_screen/Screen1View.hpp
@@ -13,7 +13,8 @@ public:
     virtual void tearDownScreen();
     virtual void LED1_button_clicked();
 
-    void updateVal(uint16_t newValue);
+    void updateVal(unsigned int newValue);
+
 protected:
 };
